+++ src/components/Multiplayer/GameLounge.tsx
@@ -10,6 +10,7 @@ import config from '../../aws-exports';
 import { createGame, updateGame } from '../../graphql/mutations';
 import { Game } from '../../models/game';
 import { authState, loginUserReducer } from '../../state-slices/auth/auth-slice';
+import { createAnswers } from './Answers'
 
 Amplify.configure(config);
 
@@ -48,7 +49,7 @@ function GameLounge() {
                     id: '10',
                     question: 'What is the answer to this question?',
                     correctAnswer: "There isn't one",
-                    multiAnswers: ['wrong', 'correct', 'idk']
+                    multiAnswers: createAnswers()
                 }]
             },
             players: [{
